services/trade-simulator/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
redis==5.0.1
structlog==24.1.0
pyjwt==2.8.0
requests==2.31.0
python-dotenv==1.0.0
marshmallow==3.20.1
